refactor(ios): split extension settings subviews
Extract palette components, the custom palette editor, and custom layout views out of the iOS extension SettingsView so the main settings entrypoint stays focused on section wiring.

Also update ERICK-141/ERICK-144 progress notes and the AI guide to reflect the reduced SettingsView surface.
